+++ packages/sync-server/app/localisation.js
@@ -201,7 +201,7 @@ const rootLocalisationSchema = yup
     previewUvciFormat: yup
       .string()
       .required()
-      .oneOf(['eudcc', 'icao']),
+      .oneOf(['tamanu', 'eudcc', 'icao']),
     covidVaccines: yup
       .array()
       .of(yup.string())